requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
lightning-utilities==0.8.0
lit==16.0.5
Markdown==3.4.3
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
Pillow==9.3.0
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytorch-lightning==2.0.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.4
yarl==1.9.2
